+++ module/Formation/src/Formation/Service/Url/UrlService.php
@@ -5,6 +5,7 @@ namespace Formation\Service\Url;
 use Fichier\Service\Fichier\FichierStorageServiceAwareTrait;
 use Fichier\Service\Storage\Adapter\Exception\StorageAdapterException;
 use Formation\Entity\Db\Inscription;
+use Formation\Entity\Db\Session;
 use Formation\Provider\NatureFichier\NatureFichier;
 use Structure\Entity\Db\Etablissement;
 use Structure\Service\StructureDocument\StructureDocumentServiceAwareTrait;
@@ -43,11 +44,22 @@ class UrlService extends \Application\Service\Url\UrlService
     /** @noinspection PhpUnused */
     public function getUrlEnquete(): string
     {
-        if (!isset($this->vars['inscription'])) throw new RuntimeException("Variable inscription non fournie à [".UrlService::class."]");
+        if (!isset($this->variables['inscription'])) throw new RuntimeException("Variable inscription non fournie à [".UrlService::class."]");
         /** @var Inscription $inscription */
-        $inscription = $this->vars['inscription'];
+        $inscription = $this->variables['inscription'];
 
-        $url = $this->fromRoute('formation/enquete/repondre-questions', ['inscription' => $inscription->getId()], [], true);
+        $url = $this->fromRoute('formation/enquete/repondre-questions', ['inscription' => $inscription->getId()], ['force_canonical' => true], true);
         return "<a href='".$url."'>Formulaire d'enquête de retour de formation</a>";
+    }
+
+    /** @noinspection PhpUnused */
+    public function getUrlListeInscritsSessionFormation(): string
+    {
+        if (!isset($this->variables['session'])) throw new RuntimeException("Variable session non fournie à [".UrlService::class."]");
+        /** @var Session $session */
+        $session = $this->variables['session'];
+
+        $url = $this->fromRoute('formation/session/generer-emargements', ['session' => $session->getId()], ['force_canonical' => true] , true);
+        return "<a href='".$url."'>Liste des inscrits à cette session de formation</a>";
     }
 }